move/rename some files and modify integration's README
